requirements.txt
streamlit
yfinance
pandas
plotly
google-genai
bs4
requests
feedparser
